Make http server an object
should be able to do

sevala.yamba() instead of passing the server object into a static function.
